# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RelayDock API
  version: 2.0.0
  summary: Authorized-credential AI gateway and control plane
  description: |
    RelayDock exposes an OpenAI-compatible data plane and a separate control
    plane. OpenAI-compatible request and response bodies are intentionally open
    to additive fields. RelayDock does not expose account registration,
    CAPTCHA, consumer session, promotion, or policy-evasion APIs.
    Every response includes X-Request-Id and X-RelayDock-Request-Id; both carry
    RelayDock's trusted, server-generated request ID.
  license:
    name: License not yet selected
servers:
  - url: http://127.0.0.1:8081
    description: Local control plane
...
paths:
...
  /v1/models:
    get:
      tags: [Gateway]
      summary: List enabled RelayDock model aliases
      operationId: listModels
      servers:
        - url: http://127.0.0.1:8080
      security:
        - relaydockApiKey: []
      responses:
        '200':
          description: Model alias list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelList'
        '401':
          $ref: '#/components/responses/GatewayUnauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/responses:
    post:
      tags: [Gateway]
      summary: Create an OpenAI-compatible response
      operationId: createResponse
      servers:
        - url: http://127.0.0.1:8080
      security:
        - relaydockApiKey: []
      parameters:
        - $ref: '#/components/parameters/ClientRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelRequest'
      responses:
        '200':
          description: OpenAI-compatible JSON response or SSE stream
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/GatewayUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/ProviderError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/chat/completions:
    post:
      tags: [Gateway]
      summary: Create an OpenAI-compatible chat completion
      operationId: createChatCompletion
      servers:
        - url: http://127.0.0.1:8080
      security:
        - relaydockApiKey: []
      parameters:
        - $ref: '#/components/parameters/ClientRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelRequest'
      responses:
        '200':
          description: OpenAI-compatible JSON completion or SSE stream
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/GatewayUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/ProviderError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/embeddings:
    post:
      tags: [Gateway]
      summary: Create OpenAI-compatible embeddings
      operationId: createEmbedding
      servers:
        - url: http://127.0.0.1:8080
      security:
        - relaydockApiKey: []
      parameters:
        - $ref: '#/components/parameters/ClientRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelRequest'
      responses:
        '200':
          description: OpenAI-compatible embedding list
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/GatewayUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/ProviderError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /api/admin/credentials/{id}/status:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    patch:
      tags: [Admin]
      summary: Enable or disable a provider credential
      operationId: setCredentialStatus
      security:
        - controlBearer: []
        - adminSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActiveDisabledPatch'
      responses:
        '200':
          description: New state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResult'
...
components:
  securitySchemes:
    relaydockApiKey:
      type: http
      scheme: bearer
      bearerFormat: rdk_live_... or rdk_test_...
      description: Downstream RelayDock API key; never use an upstream provider key here.
    controlBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: relayedock_admin_session
      description: Administrator UI access-session cookie.
    adminRefreshCookie:
      type: apiKey
      in: cookie
      name: relayedock_admin_refresh
      description: Administrator UI refresh-session cookie.
    sharedSessionCookie:
      type: apiKey
      in: cookie
      name: relayedock_session
      description: Shared control-plane access-session cookie.
    sharedRefreshCookie:
      type: apiKey
      in: cookie
      name: relayedock_refresh
      description: Shared control-plane refresh-session cookie.
    consoleSessionCookie:
      type: apiKey
      in: cookie
      name: relayedock_console_session
      description: User Console access-session cookie.
    consoleRefreshCookie:
      type: apiKey
      in: cookie
      name: relayedock_console_refresh
      description: User Console refresh-session cookie.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ClientRequestId:
      name: X-Client-Request-Id
      in: header
      description: Optional bounded caller correlation ID; RelayDock still returns its own ID in X-Request-Id and X-RelayDock-Request-Id.
      schema:
        type: string
        maxLength: 512
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required double-submit token matching the applicable relayedock_csrf, relayedock_admin_csrf, or relayedock_console_csrf cookie for state-changing cookie-authenticated requests; Bearer authentication does not require CSRF proof.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    GatewayUnauthorized:
      description: RelayDock API-key authentication failed
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Request is not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Resource or model route not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooLarge:
      description: Request body exceeds the configured limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: RelayDock API-key limit reached
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit-Requests:
          schema:
            type: integer
        X-RateLimit-Remaining-Requests:
          schema:
            type: integer
        X-RateLimit-Limit-Tokens:
          schema:
            type: integer
        X-RateLimit-Remaining-Tokens:
          schema:
            type: integer
        X-RelayDock-RateLimit-Limit-Requests:
          schema:
            type: integer
        X-RelayDock-RateLimit-Remaining-Requests:
          schema:
            type: integer
        X-RelayDock-RateLimit-Limit-Tokens:
          schema:
            type: integer
        X-RelayDock-RateLimit-Remaining-Tokens:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ProviderError:
      description: Upstream provider connection or service error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unavailable:
      description: Required state or an eligible credential is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Internal service failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code]
          properties:
            message:
              type: string
            type:
              type: string
              const: relaydock_error
            param:
              type: [string, 'null']
            code:
              type: string
...
    ModelRequest:
      type: object
      required: [model]
      properties:
        model:
          type: string
          minLength: 1
        stream:
          type: boolean
      additionalProperties: true
    GatewayModel:
      type: object
      required: [id, object, created, owned_by]
      properties:
        id:
          type: string
        object:
          type: string
          const: model
        created:
          type: integer
          format: int64
        owned_by:
          type: string
          const: relaydock
    ModelList:
      type: object
      required: [object, data]
      properties:
        object:
          type: string
          const: list
        data:
          type: array
          items:
            $ref: '#/components/schemas/GatewayModel'
...
    ActiveDisabledPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACTIVE, DISABLED]
...
    StatusResult:
      type: object
      required: [status]
      properties:
        status:
          type: string